--- src/components/Header/Logo.tsx
@@ -1,11 +1,0 @@
-import { LogoLink, LogoImage } from "../../styled/Header/LogoStyles";
-
-const Logo: React.FC = () => {
-    return (
-        <LogoLink to="/">
-            <LogoImage src="/logo.svg" alt="ReadyMe logo" />
-        </LogoLink>
-    );
-};
-
-export default Logo;
