+++ src/components/header/authBtn/AuthBtnStyles.ts
